requirements.txt
flask==3.0.3
flask-socketio==5.3.7
python-socketio==5.11.4
python-engineio==4.9.1
eventlet==0.36.1
gunicorn==22.0.0
pycryptodome==3.20.0
argon2-cffi==23.1.0
